// src/features/editableProfileCard/model/consts/consts.ts
export enum ValidateProfileError {
    INCORRECT_USER_FIRSTNAME = 'INCORRECT_USER_FIRSTNAME',
    INCORRECT_USER_LASTNAME = 'INCORRECT_USER_LASTNAME',
    INCORRECT_USER_COUNTRY = 'INCORRECT_USER_COUNTRY',
    INCORRECT_USER_AGE = 'INCORRECT_USER_AGE',
    NO_DATA = 'NO_DATA',
    SERVER_ERROR = 'SERVER_ERROR'
}